requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-login==0.6.3

# Database
supabase==2.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# Visualization
plotly==5.18.0

# Environment & Security
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0

# Utils
requests==2.31.0
python-dateutil==2.8.2

# Server
gunicorn==21.2.0
waitress==2.1.2